added journey model

+++ db/migrate/20180116192406_add_rating_to_bars.rb
@@ -1,5 +1,5 @@
 class AddRatingToBars < ActiveRecord::Migration[5.1]
   def change
-    add_column :bars, :rating, :integer
+    add_column :bars, :rating, :float
   end
 end
